# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Payment Gateway API",
        "description": "Payment Gateway API for processing payments, managing refunds, and handling webhooks",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current Host"
        }
    ],
    "paths": {
        "/api/apps": {
            "get": {
                "tags": [
                    "Apps"
                ],
                "summary": "List all apps",
                "description": "List all apps belonging to the authenticated merchant.",
                "operationId": "5f123a8d32dc3e5b9403bc19aac48205",
                "responses": {
                    "200": {
                        "description": "List of apps retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "apps": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Apps"
                ],
                "summary": "Create a new app",
                "description": "Create a new app for the authenticated merchant.",
                "operationId": "5eeae569d2320af41d900cc34c1ab0a0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Mobile Storefront"
                                    },
                                    "webhook_url": {
                                        "type": "string",
                                        "example": "https://example.com/webhook"
                                    },
                                    "confirmation_url": {
                                        "type": "string",
                                        "example": "https://example.com/confirm"
                                    },
                                    "metadata": {
                                        "type": "object",
                                        "example": {
                                            "platform": "ios"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "App created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "app": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/apps/{appId}": {
            "get": {
                "tags": [
                    "Apps"
                ],
                "summary": "Show app details",
                "description": "Show a single app (must belong to merchant).",
                "operationId": "e7ede94aefa5479431878d6a69026787",
                "parameters": [
                    {
                        "name": "appId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "App details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "app": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "App not found"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Apps"
                ],
                "summary": "Update app settings",
                "description": "Update app settings.",
                "operationId": "25ca726f11c7e02250dc58ec7a22c620",
                "parameters": [
                    {
                        "name": "appId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Web Storefront"
                                    },
                                    "webhook_url": {
                                        "type": "string",
                                        "example": "https://example.com/new-webhook"
                                    },
                                    "confirmation_url": {
                                        "type": "string",
                                        "example": "https://example.com/new-confirm"
                                    },
                                    "metadata": {
                                        "type": "object",
                                        "example": {
                                            "platform": "web"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "App updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "app": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "App not found"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Apps"
                ],
                "summary": "Suspend an app",
                "description": "Suspend / deactivate an app (soft-disable without deletion).",
                "operationId": "4711b28f3587e7f4ff719d2c1da394ee",
                "parameters": [
                    {
                        "name": "appId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "App suspended successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "App not found"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/apps/{appId}/users": {
            "get": {
                "tags": [
                    "Apps"
                ],
                "summary": "List app users",
                "description": "List users registered under a specific app.",
                "operationId": "09ed55cfa2534e0e932ee5072404914c",
                "parameters": [
                    {
                        "name": "appId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of users retrieved successfully"
                    },
                    "404": {
                        "description": "App not found"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/apps/{appId}/users/{externalUserId}": {
            "get": {
                "tags": [
                    "Apps"
                ],
                "summary": "Show app user details",
                "description": "Show a single app user by external_user_id.",
                "operationId": "2951ef27719ca3ad95f0e446a6a01820",
                "parameters": [
                    {
                        "name": "appId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "externalUserId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "App or user not found"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiKey": {
                "type": "http",
                "bearerFormat": "API Key",
                "scheme": "bearer"
            }
        }
    },
...
}